+++ mail/mail-common/src/errors/unexpected.rs
@@ -132,8 +132,8 @@ impl From<QueuedError> for Unexpected {
             // TODO: Check with Leander if there is a better Category
             QueuedError::ActionNotFound(_id) => Self::Internal,
             // TODO: Check with Leander if there is a better Category
-            QueuedError::Busy => Self::Internal,
             QueuedError::Context(_) => Self::Internal,
+            QueuedError::WorkerChannel => Self::Internal,
         }
     }
 }